Split libertem.io.dataset.base into modules
Externally, all imports should stay the same, this is just to split
up the very long base.py
