+++ src/libertem/io/dataset/base/__init__.py
@@ -1,0 +1,16 @@
+from .exceptions import DataSetException
+from .meta import DataSetMeta, PartitionStructure
+from .roi import _roi_to_indices, _roi_to_nd_indices
+from .dataset import DataSet, WritableDataSet
+from .partition import Partition, WritablePartition
+from .datatile import DataTile
+from .part3d import Partition3D, File3D, FileSet3D
+from .utils import FileTree
+
+__all__ = [
+    DataSetException, DataSetMeta, PartitionStructure,
+    _roi_to_nd_indices, _roi_to_indices,
+    DataSet, WritableDataSet, Partition, WritablePartition,
+    DataTile, Partition3D, File3D, FileSet3D,
+    FileTree,
+]
