+++ src/libertem/io/dataset/base/exceptions.py
@@ -1,0 +1,2 @@
+class DataSetException(Exception):
+    pass
